needs more cow bell
------

partition() made, not tested. TODO unPartition() --> for testing.

+++ src/Partition.java
@@ -2,12 +2,12 @@
 
 public class Partition {
 
-	private long Is;
-	private long Il;
-	private long Js;
-	private long Jl;
+	private int Is;
+	private int Il;
+	private int Js;
+	private int Jl;
 	
-	public Partition(long I, long J) {
+	public Partition(int I, int J) {
 		Il = Encoder.ceil((double)I/J);
 		Is = Encoder.floor((double)I/J);
 		Jl = I - (Is*J);
@@ -16,8 +16,8 @@ public class Partition {
 	
 	// Metodo mais estupido da minha vida
 	/*
-	public static Partition partition(long I, long J){
-		long Is, Il, Js, Jl;
+	public static Partition partition(int I, int J){
+		int Is, Il, Js, Jl;
 		
 		Il = Encoder.ceil(I/J);
 		Is = Encoder.floor(I/J);
@@ -27,35 +27,35 @@ public class Partition {
 		return new Partition(Is, Il, Js, Jl);
 	}*/
 
-	public long getIs() {
+	public int getIs() {
 		return Is;
 	}
 
-	public void setIs(long is) {
+	public void setIs(int is) {
 		Is = is;
 	}
 
-	public long getIl() {
+	public int getIl() {
 		return Il;
 	}
 
-	public void setIl(long il) {
+	public void setIl(int il) {
 		Il = il;
 	}
 
-	public long getJs() {
+	public int getJs() {
 		return Js;
 	}
 
-	public void setJs(long js) {
+	public void setJs(int js) {
 		Js = js;
 	}
 
-	public long getJl() {
+	public int getJl() {
 		return Jl;
 	}
 
-	public void setJl(long jl) {
+	public void setJl(int jl) {
 		Jl = jl;
 	}
